fix(subboard_dsp): 对齐控制消息类型解析
WHAT:
- 去掉子板 DSP control plane 中对 CONTROL_GET_TYPE 的多余 uint8 截断
- 让控制消息类型解析保持与协议宏返回类型一致

WHY:
- CONTROL_GET_TYPE 在协议头里本身返回 uint32_t 掩码结果，额外转成 uint8_t 没有必要
- 子板控制面这里的写法与其他 demo 的控制协议解析实现不一致，容易制造无意义的类型收窄
- 虽然当前消息类型值较小，但保持与协议定义一致更稳妥，也更利于后续统一维护

HOW:
- 直接使用 CONTROL_GET_TYPE(msg) 的原始返回值写入本地 uint32_t type
- 在 gimbal_node 配置下重新构建 s300_subboard_mm_consumer_demo 和 img_s300_subboard_mm_consumer_demo 验证行为未受影响

+++ Projects/Demo/Subboard_MM_Consumer_Demo/Src/subboard_dsp_control_plane.c
@@ -79,7 +79,7 @@ void subboard_dsp_control_plane_handle_control_message(uint32_t msg,
     }
 
     clear_effect(out_effect);
-    type = (uint8_t)CONTROL_GET_TYPE(msg);
+    type = CONTROL_GET_TYPE(msg);
 
     if ((type != CONTROL_MSG_TYPE_SYS) &&
         (type != CONTROL_MSG_TYPE_ACK) &&
